Feature/salary management

+++ Services/Services/PayrollService.cs
@@ -1,4 +1,7 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.SignalR;
+using Microsoft.EntityFrameworkCore;
 using Models.DTO.Request;
 using Models.Enums;
 using Models.Models;
@@ -12,11 +15,11 @@ namespace Services.Services
     {
         private readonly IMapper _mapper;
         private readonly EmsContext _context;
-        private readonly ISalaryCalculatorService _salaryCalculatorService;
+        private readonly SalaryCalculatorService _salaryCalculatorService;
         public PayrollService(
             IMapper mapper,
-            EmsContext context, 
-            ISalaryCalculatorService salaryCalculatorService)
+            EmsContext context,
+            SalaryCalculatorService salaryCalculatorService)
         {
             _mapper = mapper;
             _context = context;
@@ -25,33 +28,67 @@ namespace Services.Services
 
         public void CreatePayroll(PayrollDTO payrollDTO)
         {
-            _salaryCalculatorService.CalculateSalaryAndWritePayslipsIntoDatabase(payrollDTO);
+            _salaryCalculatorService.CalculateUserSalaryAndWriteIntoDatabases(payrollDTO);
+        }
+
+        public void SendPayroll(int payrollId)
+        {
+            var payroll = _context.Payrolls.Find(payrollId);
+            if (payroll == null)
+            {
+                throw new Exception("Payroll not found");
+            }
+
+            payroll.Status = PayrollStatus.Sent;
+            _context.Payrolls.Update(payroll);
+            _context.SaveChanges();
+        }
+
+        public List<Payslip> GetPayslipsOfPayroll(int payrollId)
+        {
+            var payroll = _context.Payrolls.Find(payrollId);
+            if (payroll == null)
+            {
+                throw new Exception("Payroll is null");
+            }
+
+            _context.Entry(payroll).Collection(p => p.PayslipList).Load();
+
+            foreach (var payslip in payroll.PayslipList)
+            {
+                _context.Entry(payslip).Reference(p => p.Employee).Load();
+            }
+
+            if (payroll.PayslipList == null)
+            {
+                return new List<Payslip>();
+            }
+
+            return payroll.PayslipList;
         }
         public void DeletePayroll(int id)
         {
-            var payroll = _context.Payrolls.Find(id);
+            var payroll = _context.Payrolls.Where(payroll => payroll.Id == id)
+                .Include(payroll => payroll.PayslipList)
+                .Single();
 
             if (payroll == null)
             {
                 throw new Exception("Cannot delete payroll of null");
             }
 
-            _context.Payrolls.Remove(payroll);
-            _context.SaveChanges();
-        }
-
-        public void SendPayroll(int id)
-        {
-            var payroll = _context.Payrolls.Find(id);
-
-            if (payroll == null)
+            if (payroll.PayslipList == null)
             {
-                throw new Exception("Pay roll cannot be null");
+                throw new Exception("Payroll not init properly");
             }
 
-            payroll.Status = PayrollStatus.Sent;
+            foreach (var payslip in payroll.PayslipList)
+            {
+                _context.Entry(payslip).Collection(p => p.Timekeepings).Load();
+                _context.Entry(payslip).Collection(p => p.SalaryDeltas).Load();
+            }
 
-            _context.Payrolls.Update(payroll);
+            _context.Payrolls.Remove(payroll);
             _context.SaveChanges();
         }
 
@@ -121,9 +158,10 @@ namespace Services.Services
                 return new List<Payslip>();
             }
 
-            return payroll.PayslipList
-                .Where(payslip => payslip.Employee.Name.Contains(query) ||
-                    query.Contains(payslip.Employee.Name))
+
+            return _context.Payslips
+                .Where(payslip => payslip.PayrollId == id)
+                .Include(p => p.Employee)
                 .Skip(offset)
                 .Take(limit)
                 .ToList();
@@ -140,5 +178,70 @@ namespace Services.Services
             return payslipList.Count();
         }
 
+        public List<PayslipWorkingShiftTimekeeping> GetPayslipTimekeepings(int id)
+        {
+            var payslip = _context.Payslips.Where(x => x.Id == id)
+                .Include(p => p.Timekeepings)
+                .FirstOrDefault();
+
+            if (payslip == null)
+            {
+                throw new Exception("Payslip not found");
+            }
+
+            if (payslip.Timekeepings == null)
+            {
+                throw new Exception("Timekeepings not initialized");
+            }
+
+            return payslip.Timekeepings;
+
+        }
+        public List<Payslip> GetPayslipsOfUser(int userId)
+        {
+            var user = _context.Users.Find(userId);
+
+            if (user == null)
+            {
+                throw new Exception("User not found");
+            }
+
+            return _context.Payslips.Where(p => p.EmployeeId == userId)
+                .Where(p => p.Payroll.Status == PayrollStatus.Sent)
+                .ToList();
+        }
+
+        public Payslip GetPayslipById(int id)
+        {
+            var payslip = _context.Payslips.Where(p => p.Id == id)
+                .Include(p => p.Employee)
+                .Single();
+
+            if (payslip == null)
+            {
+                throw new Exception("Payslip not found");
+            }
+
+            return payslip;
+        }
+
+        public List<PayslipSalaryDelta> GetPayslipSalaryDeltas(int id)
+        {
+            var payslip = _context.Payslips.Where(x => x.Id == id)
+                .Include(p => p.SalaryDeltas)
+                .FirstOrDefault();
+
+            if (payslip == null)
+            {
+                throw new Exception("Payslip not found");
+            }
+
+            if (payslip.SalaryDeltas == null)
+            {
+                throw new Exception("SalaryDeltas not initialized");
+            }
+
+            return payslip.SalaryDeltas;
+        }
     }
 }
